refactor: Refactor TasksScreen widget to use StatefulWidget and manage tasks list
This commit refactors the `TasksScreen` widget to use `StatefulWidget` instead of `StatelessWidget` and introduces a `tasks` list to manage the tasks. The `TasksScreen` widget now maintains a list of `Task` objects and displays them in a `ListView.builder`. Each `Task` object has a name and a boolean flag `isDone` to track the completion status.

The `TaskList` widget is updated to use the `tasks` list passed from the `TasksScreen` widget and dynamically generate `TaskTile` widgets based on the number of tasks in the list. The `TaskTile` widget is also modified to display the task name and a checkbox to toggle the completion status.

This refactoring improves the maintainability and extensibility of the code by separating the task management logic from the UI rendering. It also allows for easy addition of new tasks and updating the completion status.